# backend/src/main/java/com/sebsach/electronicwebshop/product/ProductResponse.java
package com.sebsach.electronicwebshop.product;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ProductResponse {
    private Long id;
    private String name;
    private String description;
    private double price;
    private String image;
    private int quantity;
    private String producer;
}